Moved polkadto wallet to tests folder
