+++ tests/polkadot_wallet/content.js
