+++ tests/polkadot_wallet/detectOther.d.ts
