+++ tests/polkadot_wallet/detectPackage.d.ts
