+++ tests/polkadot_wallet/favicon.ico
