+++ tests/polkadot_wallet/images/icon-128.png
