+++ tests/polkadot_wallet/images/icon-16.png
