+++ tests/polkadot_wallet/images/icon-32.png
