+++ tests/polkadot_wallet/images/icon-48.png
