+++ tests/polkadot_wallet/images/icon-64.png
