+++ tests/polkadot_wallet/locales/bn/translation.json
