+++ tests/polkadot_wallet/locales/fr/translation.json
