+++ tests/polkadot_wallet/locales/th/translation.json
