+++ tests/polkadot_wallet/locales/tr/translation.json
