+++ tests/polkadot_wallet/locales/zh/translation.json
