+++ tests/polkadot_wallet/manifest.json
