+++ tests/polkadot_wallet/packageInfo.d.ts
