+++ tests/polkadot_wallet/page.d.ts
